==> 2_generate_plots_GPD.R <==
##############################################################################
# Plots of grams of alcohol per day - DRINKERS
# This script generates takes the data generated by MAIHDA and generates descriptive plots of the results
##############################################################################

######################################################################## Set-up
setwd("C:/Users/Documents/")
code <- "intersectionality/MAIHDA alcohol/"
inputs <- "intersectionality/MAIHDA alcohol/inputs/"
models <- "intersectionality/MAIHDA alcohol/models/"
outputs <- "intersectionality/MAIHDA alcohol/outputs/"

library(tidyverse)
library(readr)
library(tidyr)
library(dplyr)
library(labelled)
library(haven)
library(ggplot2)
library(forcats)

## Bias toward non-scientific notation
options(scipen=10)

# Set default theme for plots:
theme_set(theme_bw(base_size = 12))

##### Plots

# Read in data
plot_data <- readRDS(paste0(outputs,"grams drinkers/results_grams_drinkers.rds"))

# Prep data for plotting
plot_data <- plot_data %>%
  mutate(sex=ifelse(grepl("Female",intersectional_names),"Women","Men"),
         race=ifelse(grepl("Asian",intersectional_names),"Asian",
                     ifelse(grepl("Black", intersectional_names),"Black",
                            ifelse(grepl("AI/AN", intersectional_names), "AI/AN",
                                ifelse(grepl("Multiple race", intersectional_names), "Multiple race",
                                   ifelse(grepl("Hispanic", intersectional_names), "Hispanic","White"))))),
         race=factor(race, levels=c("Hispanic", "AI/AN", "Asian","Black", "Multiple race", "White")),
         education=ifelse(grepl("high school", intersectional_names), "high school or less",
                          ifelse(grepl("some college", intersectional_names), "some college", "college+")),
         education=factor(education,levels=c("high school or less", "some college", "college+")),
         age=ifelse(grepl("21-24", intersectional_names), "21-24",
                    ifelse(grepl("25-59", intersectional_names), "25-59", "60+")))

# Plot of additive versus total estimates

# Create offsets for the x-axis positions
plot_data <- plot_data %>%
  mutate(education_men_add = ifelse(sex == "Men", as.numeric(education) - 0.15, as.numeric(education)),
         education_men_total = ifelse(sex == "Men", as.numeric(education) + 0.15, as.numeric(education)),
         education_women_add = ifelse(sex == "Women", as.numeric(education) - 0.15, as.numeric(education)),
         education_women_total = ifelse(sex == "Women", as.numeric(education) + 0.15, as.numeric(education)))

combined_plot <- plot_data %>%
  ggplot() +
  geom_point(data = filter(plot_data, sex == "Men"), aes(x = education_men_add, y = estAmn, color = "Men, additive effects only"), alpha = 0.5) +
  geom_errorbar(data = filter(plot_data, sex == "Men"), aes(x = education_men_add, ymin = estAlo, ymax = estAhi, color = "Men, additive effects only"), width = 0.3) +
  geom_point(data = filter(plot_data, sex == "Men"), aes(x = education_men_total, y = estmn, color = "Men, total effects")) +
  geom_errorbar(data = filter(plot_data, sex == "Men"), aes(x = education_men_total, ymin = estlo, ymax = esthi, color = "Men, total effects"), width = 0.3) +
  geom_point(data = filter(plot_data, sex == "Women"), aes(x = education_women_add, y = estAmn, color = "Women, additive effects only"), alpha = 0.5) +
  geom_errorbar(data = filter(plot_data, sex == "Women"), aes(x = education_women_add, ymin = estAlo, ymax = estAhi, color = "Women, additive effects only"), width = 0.3) +
  geom_point(data = filter(plot_data, sex == "Women"), aes(x = education_women_total, y = estmn, color = "Women, total effects")) +
  geom_errorbar(data = filter(plot_data, sex == "Women"), aes(x = education_women_total, ymin = estlo, ymax = esthi, color = "Women, total effects"), width = 0.3) +
  facet_grid(cols = vars(race), rows = vars(age)) +
  theme(axis.title.x = element_blank(), 
        axis.text.x = element_text(angle = 90, size = 12, hjust = 0.5, vjust = 0.5),
        legend.position = "bottom",
        legend.text = element_text(size = 12),
        strip.text = element_text(size = 12),
        plot.title = element_text(size = 16)) +
  theme(strip.background = element_rect(fill = "lightblue")) +
  # ggtitle("The influence of interaction effects on estimated alcohol consumption") +
  labs(y = "Estimated grams per day", color = "Sex") +
  scale_color_manual(values = c("Men, additive effects only" = "lightblue", "Men, total effects" = "darkblue", "Women, additive effects only" = "orange", "Women, total effects" = "darkred")) +
  guides(color = guide_legend(title = NULL, ncol = 4)) +
  scale_x_continuous(breaks = c(1, 2, 3), labels = c("low", "med.", "high")) +
  scale_y_continuous(breaks = seq(0, max(plot_data$esthi), by = 5))

combined_plot
ggsave(paste0(outputs,"grams drinkers/plot grams both sexes, additive versus total estimates.png"), dpi=1200, width=33, height=19, units="cm")

# Calculate the number of strata for whom there are significant interaction effects (i.e. CIs don't cross zero)
significant_interactions <- plot_data %>% filter(estIlo<0 & estIhi <0 | estIlo >0 & estIhi>0)
